src/test/java/com/kibong/refactoring_study/_04_long_parameter_list/_14_replace_parameter_with_query/OrderTest.java
package com.kibong.refactoring_study._04_long_parameter_list._14_replace_parameter_with_query;

import com.kibong.refactoring_study.long_parameter_list._14_replace_parameter_with_query.Order;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class OrderTest {

    @Test
    void discountedPriceWithDiscountLevel2() {
        int quantity = 200;
        double price = 100;
        assertEquals(quantity * price * 0.90, new Order(quantity, price).finalPrice());
    }

    @Test
    void discountedPriceWithDiscountLevel1() {
        int quantity = 100;
        double price = 100;
        assertEquals(quantity * price * 0.95, new Order(quantity, price).finalPrice());
    }

}